Cambios para corregir todo
Ya he hecho todos los cambios, sobretodo a nivel de conexiones y versión.

He tenido que actualizar el Android Studio porque no me funcionaba bien el graddle. En eclipse no he tenido que tocar nada más allá de añadir una barra para la ruta. También tuve que cambiar el modelo participante y crear una nueva base de datos.

En android, como ya he ido haciendo el código a lo largo de la semana, no ha habido muchos cambios ni muy grandes.

+++ java/Sorteo/Sorteo/src/main/java/com/example/Sorteo/ParticipanteService.java
@@ -19,7 +19,7 @@ public class ParticipanteService {
     }
 
     // Método para obtener un participante por ID
-    public Optional<Participante> findById(int id) {
+    public Optional<Participante> findById(Long id) {
         return sorteoRepository.findById(id);
     }
 
@@ -29,7 +29,7 @@ public class ParticipanteService {
     }
 
     // Método para eliminar un participante por ID
-    public void deleteById(int id) {
+    public void deleteById(Long id) {
         sorteoRepository.deleteById(id);
     }
 }
