+++ java/Sorteo/Sorteo/src/main/java/com/example/Sorteo/SorteoRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SorteoRepository extends JpaRepository<Participante, Integer> {}
+public interface SorteoRepository extends JpaRepository<Participante, Long> {}
